refactor: user mapper
